Fix sign up bug with attribute and iterator changes
This changes the league_user's ready and set_up attributes to true. The
set_up attribute change is required for the user to access trading within
the league. Although the ready attribute change is not required, setting it to
true seems appropriate.

Before this change,  the set_up attribute was defaulting to false and
preventing the user from trading in the league.

Also, replace the hard-coded times iterator with an each iterator because
the number of exchanges will fluctuate.'